Updated files

+++ src/main/java/com/ga/igdb/controller/ReviewsController.java
@@ -1,12 +1,15 @@
 package com.ga.igdb.controller;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
-
+import com.ga.igdb.dao.GameDao;
 import com.ga.igdb.dao.ReviewsDao;
 import com.ga.igdb.model.Reviews;
 
@@ -22,8 +25,15 @@ public class ReviewsController {
 	@Autowired
 	private ReviewsDao reviewsdao;
 	
-//	@Autowired
-//	private GameDao gamedao;
+	@Autowired
+	private GameDao gamedao;
+	
+	@Autowired
+	private UserController uc;
+	
+	@Autowired
+	HttpServletRequest request;
+	
 	
 	//Get - Review Add
 	@GetMapping("/reviews/add")
@@ -33,6 +43,13 @@ public class ReviewsController {
 		
 		HomeController hc = new HomeController();
 		hc.setAppName(mv, env);
+		
+		
+		if(!uc.isUserLoggedIn())
+		{
+			mv.setViewName("home/index");
+		}
+		
 		return mv;
 	}
 	
@@ -56,8 +73,8 @@ public class ReviewsController {
 		HomeController hc = new HomeController();
 		hc.setAppName(mv, env);
 		
-//		var it = gamedao.findAll();
-//		mv.addObject("games", it);
+		var t = gamedao.findAll();
+		mv.addObject("games", t);
 		
 		return mv;
 	}
@@ -91,8 +108,13 @@ public class ReviewsController {
 		HomeController hc = new HomeController();
 		hc.setAppName(mv, env);
 		
-//		var it = gamedao.findAll();
-//		mv.addObject("games", it);
+		var it = gamedao.findAll();
+		mv.addObject("games", it);
+		
+		if(!uc.isUserLoggedIn())
+		{
+			mv.setViewName("home/index");
+		}
 		
 		return mv;
 	}
@@ -100,6 +122,16 @@ public class ReviewsController {
 	//Get - Review Delete
 	@GetMapping("/reviews/delete")
 	public String deleteReview(@RequestParam int id) {
+		HttpSession session = request.getSession();
+
+		if(!uc.isUserLoggedIn())
+		{
+			return "redirect:/";
+		}
+		else if(session.getAttribute("userRole").equals("user"))
+		{
+			return "redirect:/article/index";
+		}
 		reviewsdao.deleteById(id);
 		return "redirect:/reviews/index";
 	}
